refactor(mcp): split monolithic lib.rs into schema, format, and tools modules
Extract kajet-mcp from a single 738-line lib.rs into focused modules:
- schema.rs: 4 request structs (tool input schemas)
- format.rs: 3 formatters + 11 tests
- tools.rs: 5 MCP tools + internal_error() helper
- lib.rs: ~60 lines — module declarations, KajetMcp struct, ServerHandler, serve()

Also adds list_tags i18n keys to en.toml and pl.toml.
